NEW:     - Added clang_complete plugin.     - Improved my tab completion and mapped it to a small function. _NG

+++ bundle/neocomplcache/autoload/neocomplcache/cache.vim
@@ -1,7 +1,7 @@
 "=============================================================================
 " FILE: cache.vim
 " AUTHOR: Shougo Matsushita <Shougo.Matsu@gmail.com>
-" Last Modified: 05 Jun 2012.
+" Last Modified: 02 Oct 2012.
 " License: MIT license  {{{
 "     Permission is hereby granted, free of charge, to any person obtaining
 "     a copy of this software and associated documentation files (the
@@ -30,37 +30,52 @@ set cpo&vim
 let s:Cache = vital#of('neocomplcache').import('System.Cache')
 
 " Cache loader.
-function! neocomplcache#cache#check_cache_list(cache_dir, key, async_cache_dictionary, index_keyword_list, completion_length) "{{{
+function! neocomplcache#cache#check_cache_list(cache_dir, key, async_cache_dictionary, index_keyword_list) "{{{
   if !has_key(a:async_cache_dictionary, a:key)
     return
   endif
 
-  for cache in filter(copy(a:async_cache_dictionary[a:key]),
-        \ 'filereadable(v:val.cachename)')
-    let keyword_list = []
-    for cache in a:async_cache_dictionary[a:key]
+  let keyword_list = []
+  let cache_list = a:async_cache_dictionary[a:key]
+  for cache in cache_list
+    if filereadable(cache.cachename)
       let keyword_list +=
             \ neocomplcache#cache#load_from_cache(a:cache_dir, cache.filename)
-    endfor
+    endif
+  endfor
 
-    call neocomplcache#cache#list2index(keyword_list,
-          \ a:index_keyword_list, a:completion_length)
+  call neocomplcache#cache#list2index(keyword_list, a:index_keyword_list)
+  call filter(cache_list, '!filereadable(v:val.cachename)')
 
+  if empty(cache_list)
     " Delete from dictionary.
     call remove(a:async_cache_dictionary, a:key)
-    break
-  endfor
+  endif
 endfunction"}}}
-function! neocomplcache#cache#check_cache(cache_dir, key, async_cache_dictionary, keyword_list_dictionary, completion_length) "{{{
+function! neocomplcache#cache#check_cache(cache_dir, key, async_cache_dictionary, keyword_list_dictionary) "{{{
   " Caching.
   if !has_key(a:keyword_list_dictionary, a:key)
     let a:keyword_list_dictionary[a:key] = {}
   endif
   return neocomplcache#cache#check_cache_list(
         \ a:cache_dir, a:key, a:async_cache_dictionary,
-        \ a:keyword_list_dictionary[a:key], a:completion_length)
+        \ a:keyword_list_dictionary[a:key])
 endfunction"}}}
 function! neocomplcache#cache#load_from_cache(cache_dir, filename)"{{{
+  try
+    return eval(get(neocomplcache#cache#readfile(
+          \ a:cache_dir, a:filename), 0, '[]'))
+  catch
+    let cache_name =
+          \ neocomplcache#cache#encode_name(a:cache_dir, a:filename)
+    if filereadable(cache_name)
+      call delete(cache_name)
+    endif
+
+    return []
+  endtry
+endfunction"}}}
+function! neocomplcache#cache#load_from_cache_old(cache_dir, filename)"{{{
   try
     return map(map(neocomplcache#cache#readfile(a:cache_dir, a:filename),
           \ 'split(v:val, "|||", 1)'), '{
@@ -73,11 +88,12 @@ function! neocomplcache#cache#load_from_cache(cache_dir, filename)"{{{
     return []
   endtry
 endfunction"}}}
-function! neocomplcache#cache#index_load_from_cache(cache_dir, filename, completion_length)"{{{
+function! neocomplcache#cache#index_load_from_cache(cache_dir, filename)"{{{
   let keyword_lists = {}
 
+  let completion_length = 2
   for keyword in neocomplcache#cache#load_from_cache(a:cache_dir, a:filename)
-    let key = tolower(keyword.word[: a:completion_length-1])
+    let key = tolower(keyword.word[: completion_length-1])
     if !has_key(keyword_lists, key)
       let keyword_lists[key] = []
     endif
@@ -86,9 +102,10 @@ function! neocomplcache#cache#index_load_from_cache(cache_dir, filename, complet
 
   return keyword_lists
 endfunction"}}}
-function! neocomplcache#cache#list2index(list, dictionary, completion_length)"{{{
+function! neocomplcache#cache#list2index(list, dictionary)"{{{
+  let completion_length = 2
   for keyword in a:list
-    let key = tolower(keyword.word[: a:completion_length-1])
+    let key = tolower(keyword.word[: completion_length-1])
     if !has_key(a:dictionary, key)
       let a:dictionary[key] = {}
     endif
@@ -99,6 +116,10 @@ function! neocomplcache#cache#list2index(list, dictionary, completion_length)"{{
 endfunction"}}}
 
 function! neocomplcache#cache#save_cache(cache_dir, filename, keyword_list)"{{{
+  call neocomplcache#cache#writefile(
+        \ a:cache_dir, a:filename, [string(a:keyword_list)])
+endfunction"}}}
+function! neocomplcache#cache#save_cache_old(cache_dir, filename, keyword_list)"{{{
   " Create dictionary key.
   for keyword in a:keyword_list
     if !has_key(keyword, 'abbr')
@@ -228,9 +249,8 @@ function! neocomplcache#cache#async_load_from_tags(cache_dir, filename, filetype
     let tags_file_name =
           \ neocomplcache#cache#encode_name('tags_output', a:filename)
 
-    let args = has_key(g:neocomplcache_ctags_arguments_list, a:filetype) ?
-          \ g:neocomplcache_ctags_arguments_list[a:filetype]
-          \ : g:neocomplcache_ctags_arguments_list['default']
+    let default = get(g:neocomplcache_ctags_arguments_list, '_', '')
+    let args = get(g:neocomplcache_ctags_arguments_list, a:filetype, default)
 
     if has('win32') || has('win64')
       let filename =
@@ -277,7 +297,7 @@ function! s:async_load(argv, cache_dir, filename)"{{{
         if !executable('/Applications/MacVim.app/Contents/MacOS/Vim')
           call neocomplcache#print_error(
                 \ 'You installed MacVim in not default directory!'.
-                \ ' You must add MacVim install path in $PATH.')
+                \ ' You must add MacVim installed path in $PATH.')
           let g:neocomplcache_use_vimproc = 0
           return
         endif
@@ -285,13 +305,15 @@ function! s:async_load(argv, cache_dir, filename)"{{{
         let vim_path = '/Applications/MacVim.app/Contents/MacOS/Vim'
       else
         call neocomplcache#print_error(
-              \ printf('Vim path : "%s" is not found.', v:progname))
+              \ printf('Vim path : "%s" is not found.'.
+              \        ' You must add "%s" installed path in $PATH.',
+              \        v:progname, v:progname))
         let g:neocomplcache_use_vimproc = 0
         return
       endif
     else
       let base_path = neocomplcache#util#substitute_path_separator(
-            \ fnamemodify(vimproc#get_command_name(paths[0]), ':p:h'))
+            \ fnamemodify(paths[0], ':p:h'))
 
       let vim_path = base_path .
             \ (neocomplcache#util#is_windows() ? '/vim.exe' : '/vim')
@@ -304,7 +326,7 @@ function! s:async_load(argv, cache_dir, filename)"{{{
 
     if !executable(vim_path)
       call neocomplcache#print_error(
-            \ printf('Vim path : "%s" is not found.', vim_path))
+            \ printf('Vim path : "%s" is not executable.', vim_path))
       let g:neocomplcache_use_vimproc = 0
       return
     endif
